requirements.txt
#requirements.txt
google-cloud-bigquery
google-auth